Clean up project and add nuget info
