+++ .gitignore
@@ -397,3 +397,4 @@ FodyWeavers.xsd
 # JetBrains Rider
 *.sln.iml
 /DAS.Trader.IntegrationClient/appsettings.local.json
+/DAS.Trader.SampleApp/appsettings.local.json
