+++ DAS.Trader.IntegrationClient.sln
@@ -3,7 +3,14 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.3.32929.385
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DAS.Trader.IntegrationClient", "DAS.Trader.IntegrationClient.csproj", "{C54C7315-DE8A-460D-8570-D3E2B4E18E14}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DAS.Trader.IntegrationClient", "DAS.Trader.IntegrationClient\DAS.Trader.IntegrationClient.csproj", "{C54C7315-DE8A-460D-8570-D3E2B4E18E14}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DAS.Trader.SampleApp", "DAS.Trader.SampleApp\DAS.Trader.SampleApp.csproj", "{CC35E212-1529-4692-830D-0600E21941AE}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Docs", "Docs", "{7E3D31A2-3C1D-46D2-AA11-A26149A90D3F}"
+	ProjectSection(SolutionItems) = preProject
+		Docs\CMD API Manual.pdf = Docs\CMD API Manual.pdf
+	EndProjectSection
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +22,10 @@ Global
 		{C54C7315-DE8A-460D-8570-D3E2B4E18E14}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C54C7315-DE8A-460D-8570-D3E2B4E18E14}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C54C7315-DE8A-460D-8570-D3E2B4E18E14}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CC35E212-1529-4692-830D-0600E21941AE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CC35E212-1529-4692-830D-0600E21941AE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CC35E212-1529-4692-830D-0600E21941AE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CC35E212-1529-4692-830D-0600E21941AE}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
