+++ DAS.Trader.IntegrationClient/Commands/CommandResult.cs
@@ -1,4 +1,4 @@
-﻿using DAS.Trader.IntegrationClient.Commands.Interfaces;
+﻿using DAS.Trader.IntegrationClient.Interfaces;
 
 namespace DAS.Trader.IntegrationClient.Commands;
 
@@ -7,5 +7,5 @@ internal class CommandResult : ICommandResult
     public static ICommandResult SuccessResult => new CommandResult { Success = true };
 
     public bool Success { get; set; }
-    public string Message { get; set; }
+    public string? Message { get; set; }
 }
