--- DAS.Trader.IntegrationClient/Commands/Interfaces/ICommandResult.cs
@@ -1,7 +1,0 @@
-﻿namespace DAS.Trader.IntegrationClient.Commands.Interfaces;
-
-public interface ICommandResult
-{
-    bool Success { get; }
-    string Message { get; }
-}
