+++ DAS.Trader.IntegrationClient/Commands/TcpCommands/BaseSlOfferOperationCommand.cs
@@ -1,0 +1,13 @@
+﻿using DAS.Trader.IntegrationClient.Enums;
+
+namespace DAS.Trader.IntegrationClient.Commands.TcpCommands
+{
+    public abstract class BaseSlOfferOperationCommand : BaseTcpCommand
+    {
+        protected BaseSlOfferOperationCommand(string locateOrderId, string operation) : base(
+            TraderCommandType.SLOFFEROPERATION_COMMAND, false, false,
+            locateOrderId, operation)
+        {
+        }
+    }
+}
