+++ DAS.Trader.IntegrationClient/Commands/TcpCommands/BaseSubscribeCommand.cs
@@ -1,0 +1,13 @@
+﻿using DAS.Trader.IntegrationClient.Enums;
+
+namespace DAS.Trader.IntegrationClient.Commands.TcpCommands
+{
+    public abstract class BaseSubscribeCommand : BaseTcpCommand
+    {
+        protected BaseSubscribeCommand(string symbol, params string[] parameters) : base(
+            TraderCommandType.SB_COMMAND, false, false,
+            new[] { symbol }.Concat(parameters).ToArray())
+        {
+        }
+    }
+}
