+++ DAS.Trader.IntegrationClient/Commands/TcpCommands/SlCancelOrderCommand.cs
@@ -1,0 +1,13 @@
+﻿using DAS.Trader.IntegrationClient.Enums;
+
+namespace DAS.Trader.IntegrationClient.Commands.TcpCommands
+{
+    public sealed class SlCancelOrderCommand : BaseTcpCommand
+    {
+        public SlCancelOrderCommand(string locateOrderId) : base(
+            TraderCommandType.SLCANCELORDER_COMMAND, false, false,
+            locateOrderId)
+        {
+        }
+    }
+}
