+++ DAS.Trader.IntegrationClient/Commands/TcpCommands/SlOfferOperationAcceptCommand.cs
@@ -1,0 +1,9 @@
+﻿namespace DAS.Trader.IntegrationClient.Commands.TcpCommands
+{
+    public sealed class SlOfferOperationAcceptCommand : BaseSlOfferOperationCommand
+    {
+        public SlOfferOperationAcceptCommand(string locateOrderId) : base(locateOrderId, "Accept")
+        {
+        }
+    }
+}
