+++ DAS.Trader.IntegrationClient/Commands/TcpCommands/SlOfferOperationRejectCommand.cs
@@ -1,0 +1,9 @@
+﻿namespace DAS.Trader.IntegrationClient.Commands.TcpCommands
+{
+    public sealed class SlOfferOperationRejectCommand : BaseSlOfferOperationCommand
+    {
+        public SlOfferOperationRejectCommand(string locateOrderId) : base(locateOrderId, "Reject")
+        {
+        }
+    }
+}
