+++ DAS.Trader.IntegrationClient/Commands/TcpCommands/SubscribeTimeSalesCommand.cs
@@ -1,0 +1,9 @@
+﻿namespace DAS.Trader.IntegrationClient.Commands.TcpCommands
+{
+    public sealed class SubscribeTimeSalesCommand : BaseSubscribeCommand
+    {
+        public SubscribeTimeSalesCommand(string symbol) : base(symbol, "tms")
+        {
+        }
+    }
+}
