+++ DAS.Trader.IntegrationClient/Commands/TcpCommands/UnSubscribeDayChartCommand.cs
@@ -1,0 +1,9 @@
+﻿namespace DAS.Trader.IntegrationClient.Commands.TcpCommands
+{
+    public sealed class UnSubscribeDayChartCommand : BaseUnSubscribeCommand
+    {
+        public UnSubscribeDayChartCommand(string symbol) : base(symbol, "DAYCHART")
+        {
+        }
+    }
+}
