+++ DAS.Trader.IntegrationClient/Commands/TcpCommands/UnSubscribeMinChartCommand.cs
@@ -1,0 +1,9 @@
+﻿namespace DAS.Trader.IntegrationClient.Commands.TcpCommands
+{
+    public sealed class UnSubscribeMinChartCommand : BaseUnSubscribeCommand
+    {
+        public UnSubscribeMinChartCommand(string symbol) : base(symbol, "MINCHART")
+        {
+        }
+    }
+}
