+++ DAS.Trader.IntegrationClient/Commands/TcpCommands/UnSubscribeTimeSalesCommand.cs
@@ -1,0 +1,9 @@
+﻿namespace DAS.Trader.IntegrationClient.Commands.TcpCommands
+{
+    public sealed class UnSubscribeTimeSalesCommand : BaseUnSubscribeCommand
+    {
+        public UnSubscribeTimeSalesCommand(string symbol) : base(symbol, "tms")
+        {
+        }
+    }
+}
