+++ DAS.Trader.IntegrationClient/Common/EnumHelper.cs
@@ -1,7 +1,6 @@
 ﻿using System.Collections.Concurrent;
 using System.ComponentModel;
 using System.Globalization;
-using DAS.Trader.IntegrationClient.Commands;
 
 namespace DAS.Trader.IntegrationClient.Common;
