+++ DAS.Trader.IntegrationClient/Interfaces/ICommandResult.cs
@@ -1,0 +1,7 @@
+﻿namespace DAS.Trader.IntegrationClient.Interfaces;
+
+public interface ICommandResult
+{
+    bool Success { get; }
+    string? Message { get; }
+}
