+++ DAS.Trader.SampleApp/appsettings.json
